third_party/blink/web_tests/virtual/scoped-custom-element-registry-disabled/external/wpt/custom-elements/scoped-registry/constructor-reentry-with-different-definition.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] Re-entry via upgrade before calling super()
  Failed to construct 'CustomElementRegistry': Illegal constructor
[FAIL] Re-entry via upgrade after calling super()
  Failed to construct 'CustomElementRegistry': Illegal constructor
[FAIL] Re-entry via direct constructor call before calling super()
  Failed to construct 'CustomElementRegistry': Illegal constructor
[FAIL] Re-entry via direct constructor call after calling super()
  Failed to construct 'CustomElementRegistry': Illegal constructor
Harness: the test ran to completion.
